unsupported/office365mgmt/1.0.0/requirements.txt
# No extra requirements needed
request
